(1) removed streamlit and warning (2) got rid of notebook_name argument for cell magic (3) have a clean Xu_sepsis notebook (4) Readme (5) requirements [WIP]
