+++ chapyter/requirements.txt
@@ -1,0 +1,12 @@
+# Core Libraries
+numpy
+pandas
+
+# Visualization Libraries
+matplotlib=3.4.3
+seaborn=0.11.2
+plotly=5.3.1
+
+# Machine Learning Libraries
+scikit-learn=0.24.2
+scipy=1.7.1
